% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{gen_model_formula}
\alias{gen_model_formula}
\title{Generate the model formula for the given model type}
\usage{
gen_model_formula(model_type)
}
\description{
Generate the model formula for the given model type
}
