% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{get_params_simu}
\alias{get_params_simu}
\title{Return data simulation parameters for a given model}
\usage{
get_params_simu(model_type)
}
\description{
Return data simulation parameters for a given model
}
