% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{get_target_coeff}
\alias{get_target_coeff}
\title{Return a simulation coefficient for a given parameter}
\usage{
get_target_coeff(param, model_type)
}
\description{
Return a simulation coefficient for a given parameter
}
