% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{set_params_simu}
\alias{set_params_simu}
\title{Set data simulation parameters for a given model}
\usage{
set_params_simu(x, model_type)
}
\description{
Set data simulation parameters for a given model
}
