% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_demo_pipeline}
\alias{setup_demo_pipeline}
\title{Generate a \verb{_targets.R} that defines the analytical pipeline}
\usage{
setup_demo_pipeline()
}
\description{
Generate a \verb{_targets.R} that defines the analytical pipeline
}
