% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf-hyoung-ffcws-package.R
\docType{package}
\name{osfHYoungFFCWS-package}
\alias{osfHYoungFFCWS}
\alias{osfHYoungFFCWS-package}
\title{osfHYoungFFCWS: Reproducible Analytical Pipeline Demonstration For OSF Pre-registration}
\description{
Demonstration pipeline for the main analyses described in the OSF pre-registration 'TITLE' (LINK). It includes the simulation of data and and parameter recovery through bayesian estimation of linear models.
}
\author{
\strong{Maintainer}: Hugo Mell \email{hugo.mell@ipea7892.fr} (\href{https://orcid.org/0000-0002-2346-1008}{ORCID})

}
\keyword{internal}
